fix the router
