+++ package.json
@@ -15,9 +15,9 @@
     "firebase": "^10.12.5",
     "framer-motion": "^11.3.24",
     "mini-svg-data-uri": "^1.4.4",
-    "next": "14.2.5",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^14.2.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "tailwind-merge": "^2.5.2"
   },
   "devDependencies": {
